tests: merge placeholder_test and format_test files
